+++ src/composables/connection.ts
@@ -1,9 +1,10 @@
 import { readonly, ref } from 'vue';
 import { Connection } from '../types/connection';
 import kafkaService from '../services/kafka';
+import { defineStore } from 'pinia';
 
-export function useConnection() {
-	const connection = ref<Connection | undefined>();
+export const useConnectionStore = defineStore('connection', () => {
+	const connection = ref<Connection>();
 
 	async function setConnection(newConnection: Connection) {
 		const groupId = `${newConnection.groupPrefix ? `${newConnection.groupPrefix}.` : ''}kafka-panel`;
@@ -16,4 +17,4 @@ export function useConnection() {
 		// Override setConnection function so we can update our connection
 		setConnection,
 	};
-}
+});