+++ src/services/autosends.ts
@@ -1,10 +1,8 @@
-import { faker } from '@faker-js/faker';
-import { clone } from 'ramda';
+import { Duration } from 'luxon';
+import { BehaviorSubject } from 'rxjs';
 import { ActiveAutosend, AutosendTime } from '../types/autosend';
 import kafkaService from './kafka';
-import { BehaviorSubject } from 'rxjs';
 import { Timer } from './Timer';
-import { Duration } from 'luxon';
 
 class AutosendsService {
 	private intervals: Record<string, NodeJS.Timer> = {};
@@ -24,17 +22,15 @@ class AutosendsService {
 		const interval = this.castAutosendTimeToDuration(autosend.options.interval);
 		this.intervals[autosend.id] = setInterval((autosend: ActiveAutosend) => {
 			try {
-				const interpolatedKey = this.interpolateFakeValues(clone(autosend.key), {faker});
-				const interpolatedValue = this.interpolateFakeValues(clone(autosend.value), {faker, key: interpolatedKey});
-
-				kafkaService.sendMessage(
-					autosend.topic,
-					JSON.stringify(interpolatedKey),
-					JSON.stringify(interpolatedValue)
-				).then(() => {
-					messagesSent += 1;
-					messagesCounter.next(messagesSent);
-				});
+				kafkaService.sendMessage(autosend.topic, {
+					headers: autosend.headers,
+					key: autosend.key,
+					value: autosend.value
+				})
+					.then(() => {
+						messagesSent += 1;
+						messagesCounter.next(messagesSent);
+					});
 			} catch (error) {
 				messagesCounter.error(error);
 				this.stopAutosend(autosend);
@@ -50,35 +46,6 @@ class AutosendsService {
 
 	stopAutosend(autosend: ActiveAutosend) {
 		clearInterval(this.intervals[autosend.id]);
-	}
-
-	private interpolateFakeValues(input: unknown, context: unknown) {
-		if (typeof input === 'object') {
-			if (Array.isArray(input)) {
-				for (const key in input) {
-					input[key] = this.interpolateFakeValues(input[key], context);
-				}
-			} else if (input !== null) {
-				for (const key of Object.keys(input)) {
-					const obj = input as Record<string, unknown>;
-					obj[key] = this.interpolateFakeValues(obj[key], context);
-				}
-			}
-		} else if (typeof input === 'string') {
-			input = input.replace(/\{\{(.*?)\}\}/gm, (_, group) => {
-				try {
-					return this.scopeEval(context, `this.${group}`).toString();
-				} catch (error) {
-					throw new Error(`Interpolation error with "${group}"\n${error}`);
-				}
-			});
-		}
-
-		return input;
-	}
-
-	private scopeEval(scope: unknown, script: string) {
-		return Function('"use strict";return (' + script + ')').bind(scope)();
 	}
 
 	private castAutosendTimeToDuration(time: AutosendTime) {