+++ src/services/kafka.ts
@@ -1,27 +1,29 @@
+import { faker } from '@faker-js/faker';
+import { clone } from 'ramda';
 import { invoke } from '@tauri-apps/api';
+import { emit, listen, UnlistenFn } from '@tauri-apps/api/event';
+import { Subject } from 'rxjs';
 import { SaslConfig } from '../types/connection';
 import { ConsumerGroup, ConsumerGroupState } from '../types/consumerGroup';
-import { KafkaMessage } from '../types/message';
+import { Message, MessageContent } from '../types/message';
 import { Topic } from '../types/topic';
-import { UnlistenFn, emit, listen } from '@tauri-apps/api/event';
-import { Subject } from 'rxjs';
 
 class KafkaService {
 	private unlisten?: UnlistenFn;
 	private batchingInterval?: NodeJS.Timer;
-	private batchMessages: KafkaMessage[] = [];
+	private batchMessages: Message[] = [];
 	
 	async setConnection(brokers: string[], groupId: string, sasl?: SaslConfig) {
 		await invoke('set_connection_command', {brokers, groupId, sasl});
 	}
 
 	async getTopicsState() {
-		const topicsGroups: Record<string, ConsumerGroupState> = await invoke('get_topics_state_command');
+		const topicsGroups = await invoke<Record<string, ConsumerGroupState>>('get_topics_state_command');
 		return topicsGroups;
 	}
 
 	async listGroupsFromTopic(topicName: string) {
-		const groups: ConsumerGroup[] = await invoke('get_groups_from_topic_command', {topicName});
+		const groups = await invoke<ConsumerGroup[]>('get_groups_from_topic_command', {topicName});
 		return groups;
 	}
 
@@ -30,7 +32,7 @@ class KafkaService {
 	}
 
 	async listTopics() {
-		const topics: Topic[] = await invoke('get_topics_command');
+		const topics = await invoke<Topic[]>('get_topics_command');
 		return topics;
 	}
 
@@ -49,7 +51,7 @@ class KafkaService {
 	}
 
 	async listenMessages(topic: string, messagesNumber: number) {
-		const messagesSubject = new Subject<KafkaMessage[]>();
+		const messagesSubject = new Subject<Message[]>();
 
 		if (this.unlisten) {
 			messagesSubject.error('Unexpected error: previous listening for messages is still active');
@@ -68,7 +70,7 @@ class KafkaService {
 				messagesSubject.error(error);
 			});
 
-		const unlisten = await listen<KafkaMessage>('onMessage', (event) => {
+		const unlisten = await listen<Message>('onMessage', (event) => {
 			const kafkaMessage = event.payload;
 			this.batchMessages.push(kafkaMessage);
 		});
@@ -91,8 +93,46 @@ class KafkaService {
 		await emit('offMessage');
 	}
 
-	async sendMessage(topic: string, key: string, value: string) {
-		await invoke('send_message_command', {topic, key, value});
+	async sendMessage(topic: string, message: MessageContent) {
+		const interpolatedHeaders = this.interpolateFakeValues(clone(message.headers), {faker});
+		const interpolatedKey = this.interpolateFakeValues(clone(message.key), {faker});
+		const interpolatedValue = this.interpolateFakeValues(clone(message.value), {faker, key: interpolatedKey});
+		
+		await invoke('send_message_command', {
+			topic, 
+			headers: interpolatedHeaders,
+			key: interpolatedKey,
+			value: interpolatedValue,
+		});
+	}
+
+	private interpolateFakeValues<T = unknown>(input: T, context: Record<string, unknown>): T {
+		if (typeof input === 'object') {
+			if (Array.isArray(input)) {
+				for (const key in input) {
+					input[key] = this.interpolateFakeValues(input[key], context);
+				}
+			} else if (input !== null) {
+				for (const key of Object.keys(input)) {
+					const obj = input as Record<string, unknown>;
+					obj[key] = this.interpolateFakeValues(obj[key], context);
+				}
+			}
+		} else if (typeof input === 'string') {
+			(input as string) = input.replace(/\{\{(.*?)\}\}/gm, (_, group) => {
+				try {
+					return this.scopeEval(context, `this.${group}`).toString();
+				} catch (error) {
+					throw new Error(`Interpolation error with "${group}"\n${error}`);
+				}
+			});
+		}
+
+		return input;
+	}
+
+	private scopeEval(scope: unknown, script: string) {
+		return Function('"use strict";return (' + script + ')').bind(scope)();
 	}
 }